+++ R/hydro_imgw_daily.R
@@ -3,13 +3,8 @@
 #' Downloading daily hydrological data from the danepubliczne.imgw.pl collection
 #'
 #' @param year vector of years (e.g., 1966:2000)
-#' @param coords add coordinates of the stations (logical value TRUE or FALSE)
 #' @param station name or ID of hydrological station(s).
 #' It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)
-#' @param col_names three types of column names possible:
-#' "short" - default, values with shorten names,
-#' "full" - full English description,
-#' "polish" - original names in the dataset
 #' @param allow_failure logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE
 #' @param ... other parameters that may be passed to the 'shortening' function that shortens column names
 #' @importFrom utils download.file unzip read.csv
@@ -22,18 +17,14 @@
 #'
 
 hydro_imgw_daily = function(year,
-                            coords = FALSE,
                             station = NULL,
-                            col_names = "short",
                             allow_failure = TRUE,
                             ...) {
   
   # nocov start
   if (allow_failure) {
     tryCatch(hydro_imgw_daily_bp(year,
-                                 coords,
                                  station,
-                                 col_names, 
                                  ...),
              error = function(e){
                message(paste("Problems with downloading data.",
@@ -42,9 +33,7 @@ hydro_imgw_daily = function(year,
   # nocov end
   } else {
     hydro_imgw_daily_bp(year,
-                        coords,
                         station,
-                        col_names, 
                         ...)
   }
 }
@@ -52,9 +41,7 @@ hydro_imgw_daily = function(year,
 #' @keywords internal
 #' @noRd
 hydro_imgw_daily_bp = function(year,
-                               coords,
                                station,
-                               col_names,
                                ...) {
   translit = check_locale()
   base_url = "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_hydrologiczne/"
@@ -105,21 +92,14 @@ hydro_imgw_daily_bp = function(year,
         test_url(link = address, output = temp)
         unzip(zipfile = temp, exdir = temp2)
         file1 = paste(temp2, dir(temp2), sep = "/")[1]
-        
         data1 = imgw_read(translit, file1)
-        # extra exception for a current year according to information provided by IMGW-PIB:, i.e.:
-        # "Do czasu zakonczenia kontroli przeplywow rekordy z danymi z roku 2020 maja format:
-        # Kod stacji  #Nazwa stacji  #Nazwa rzeki/jeziora  #Rok hydrologiczny  #Wskaznik miesiaca w roku hydrologicznym
-        # Dzien  #Stan wody [cm]  #Temperatura wody [st. C]    #Miesiac kalendarzowy
+        
         if (ncol(data1) == 9) {
           data1$flow = NA
           data1 = data1[, c(1:7, 10, 8:9)]
         }
         
         colnames(data1) = meta[[1]]$parameters
-        for (labs in seq_along(meta[[1]]$parameters)) {
-          attr(data1[[labs]], "label") = meta[[1]]$label[[labs]]
-        }
         codz_data = rbind(codz_data, data1)
       } # end of codz_
       
@@ -134,9 +114,6 @@ hydro_imgw_daily_bp = function(year,
         file2 = paste(temp2, dir(temp2), sep = "/")[1]
         data2 = imgw_read(translit, file2)
         colnames(data2) = gsub(x = meta[[2]]$parameters, "^ZJ", "CO") # rename colnames starting with ^ZJ to be changed to ^CO:
-        for (labs in seq_along(meta[[2]]$parameters)) {
-          attr(data2[[labs]], "label") = meta[[2]]$label[[labs]]
-        }
         zjaw_data = rbind(zjaw_data, data2)
       }
       
@@ -169,12 +146,10 @@ hydro_imgw_daily_bp = function(year,
   all_data[all_data == 99.9] = NA            
   all_data[all_data == 999] = NA
   
-  if (coords) {
-    all_data = merge(climate::imgw_hydro_stations, all_data,
-                     by.x = "id",
-                     by.y = "PSKDSZS",
-                     all.y = TRUE)
-  }
+  all_data = merge(climate::imgw_hydro_stations, all_data,
+                   by.x = "id",
+                   by.y = "PSKDSZS",
+                   all.y = TRUE)
   
   all_data = as.data.frame(all_data)
   all_data = all_data[do.call(order, all_data[grep(x = colnames(all_data), "PSNZWP|COROKH|COMSCH|CODZIEN")]), ]
@@ -185,12 +160,20 @@ hydro_imgw_daily_bp = function(year,
   data_df = all_data[, c(yy_ind, mm_ind, dd_ind)]
   data_df$yy = ifelse(data_df[, 2] >= 11, data_df[, 1] - 1, data_df[, 1])
   all_data$Data = as.Date(ISOdate(year = data_df$yy, month = data_df[, 2], day = data_df[, 3]))
+  all_data = all_data[order(all_data$id, all_data$Data), ]
   #all_data = all_data[, c(1:3, ncol(all_data), 4:(ncol(all_data) - 1)), ]
   # zjaw parameters were renamed ZJ → CO earlier; mirror that in the label map
   meta_co = meta[[2]]
   meta_co$parameters = gsub("^ZJ", "CO", meta_co$parameters)
-  all_data = imgw_rename_params_to_labels(all_data, rbind(meta[[1]], meta_co))
-  all_data = hydro_shortening_imgw(all_data, col_names = col_names, ...)
-
+  meta = unique(data.table::rbindlist(list(rbindlist(meta), meta_co)))
+  
+  # add attributes
+  for (cols in colnames(all_data)) {
+    ind = which(cols == meta$parameters)
+    if (length(ind) > 0) {
+      attr(all_data[[cols]], "label") = meta$label[ind]
+    }
+  }
+  
   return(all_data)
 }